+++ src/index.ts
@@ -6,3 +6,14 @@ export * from './rx-stomp-rpc-config.js';
 export * from './rx-stomp-rpc.js';
 export * from './i-rx-stomp-publish-params.js';
 export * from './i-watch-params.js';
+
+export {
+  IMessage,
+  IFrame,
+  IPublishParams,
+  StompHeaders,
+  IStompSocket,
+  Versions,
+  StompSocketState,
+  ITransaction,
+} from '@stomp/stompjs';