+++ src/rx-stomp-rpc.ts
@@ -5,7 +5,10 @@ import { v4 as uuid } from 'uuid';
 import { IMessage, IPublishParams, StompHeaders } from '@stomp/stompjs';
 
 import { RxStomp } from './rx-stomp.js';
-import { RxStompRPCConfig, setupReplyQueueFnType } from './rx-stomp-rpc-config.js';
+import {
+  RxStompRPCConfig,
+  setupReplyQueueFnType,
+} from './rx-stomp-rpc-config.js';
 
 /**
  * An implementation of Remote Procedure Call (RPC) using messaging.